// utils/enums.ts
export const enum SeasonName {
  // these values are in lower case for consistency
  SEASON_1 = 'season_1',
  SEASON_2 = 'season_2',
  SEASON_3 = 'season_3',
}

export const enum AssetType {
  // these values are in lower case for consistency
  LAND = 'land',
  BUILDING = 'building',
  OBJECT = 'object',
  PARTNER = 'partner',
}

export const enum AssetStatus {
  // these values are in lower case for consistency
  AVAILABLE = 'available',
  OWNED = 'owned',
  DISCARD = 'discard',
}

export const enum Provider {
  // these values are in lower case for consistency
  Wallet = 'wallet',
  Google = 'google',
  Microsoft = 'microsoft',
  Facebook = 'facebook',
  Apple = 'apple',
}

export const enum InventoryType {
  // these values are in lower case for consistency
  REWARD = 'reward',
}

export const enum UserType {
  // these values are in lower case for consistency
  ADMIN = 'admin',
  USER = 'user',
  PUBLIC = 'public',
}